Streaming tjeneste ChilFlix
In the project, me and my group have developed a software system as a streaming services such as Netflix, HBO, Viaplay etc. In the system there are media that can be played, movies or series. Both are divided into categories and series also consist of seasons and episodes. This is purely text-based. Some of the problems we had, was the path to the txt files. To make sure it run, make a new copy of the path to the txt file.